Rename fp/util.js to fp-util.js
It collides with the precompiled fp module
