+++ lib/fp/convert.js
@@ -1,6 +1,6 @@
 var _ = require('lodash'),
     baseConvert = require('./baseConvert.js'),
-    util = require('./util.js');
+    util = require('./fp-util.js');
 
 /**
  * Converts `func` of `name` to an auto-curried iteratee-first data-last version.
