+++ lib/fp/fp-util.js
